Fixed data quality unit test to actually compare values - turns out there are actually problems. 9 test cases are now failng

+++ Interfaces/Events/Docked.cs
@@ -57,6 +57,6 @@ namespace DW.ELA.LogModel.Events
         public string NameLocalised { get; set; }
 
         [JsonProperty("Proportion")]
-        public long Proportion { get; set; }
+        public double Proportion { get; set; }
     }
 }
